+++ agave_app/GLView3D.cpp
@@ -7,8 +7,9 @@
 #include "renderlib/Logging.h"
 #include "renderlib/RenderGL.h"
 #include "renderlib/RenderGLPT.h"
+#include "renderlib/gl/Image3D.h"
 #include "renderlib/gl/Util.h"
-#include "renderlib/gl/V33Image3D.h"
+
 
 #include <glm.h>
