+++ renderlib/gl/CMakeLists.txt
@@ -6,8 +6,8 @@ target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/Image2D.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Util.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Util.h"
-	"${CMAKE_CURRENT_SOURCE_DIR}/V33FSQ.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/V33FSQ.h"
-	"${CMAKE_CURRENT_SOURCE_DIR}/V33Image3D.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/V33Image3D.h"
+	"${CMAKE_CURRENT_SOURCE_DIR}/FSQ.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/FSQ.h"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Image3D.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Image3D.h"
 )
