+++ renderlib/gl/FSQ.cpp
@@ -1,4 +1,4 @@
-#include "gl/V33FSQ.h"
+#include "gl/FSQ.h"
 #include "gl/Util.h"
 
 #include <iostream>
